125B.cpp
B. Simple XML
Let's define a string <x> as an opening tag, where x is any small letter of the Latin alphabet. Each opening tag matches a closing tag of the type </x>, where x is the same letter.

Tegs can be nested into each other: in this case one opening and closing tag pair is located inside another pair.

Let's define the notion of a XML-text:

an empty string is a XML-text
if s is a XML-text, then s'=<a>+s+</a> also is a XML-text, where a is any small Latin letter
if s1, s2 are XML-texts, then s1+s2 also is a XML-text
You are given a XML-text (it is guaranteed that the text is valid), your task is to print in the following form:

each tag (opening and closing) is located on a single line
print before the tag 2 * h spaces, where h is the level of the tag's nestedness.
Input
The input data consists on the only non-empty string — the XML-text, its length does not exceed 1000 characters. It is guaranteed that the text is valid. The text contains no spaces.

Output
Print the given XML-text according to the above-given rules.

Examples
inputCopy
<a><b><c></c></b></a>
outputCopy
<a>
  <b>
    <c>
    </c>
  </b>
</a>
inputCopy
<a><b></b><d><c></c></d></a>
outputCopy
<a>
  <b>
  </b>
  <d>
    <c>
    </c>
  </d>
</a>